+++ tests/test_grouped_marginalize_combinations.py
@@ -16,7 +16,7 @@ from textwrap import dedent
 import pytest
 import torch
 
-from quivers.continuous.bayesian import marginalize_grouped
+from quivers.continuous.plate import marginalize_grouped
 
 
 _LOCAL_GRAMMAR = pytest.mark.skipif(